+++ front-end/src/components/myPage/Content.jsx
